Split exceptions to multiple modules to avoid cyclic imports
