+++ src/dishka/async_container.py
@@ -15,11 +15,11 @@ from .container_objects import Exit
 from .context_proxy import ContextProxy
 from .dependency_source import Factory
 from .entities.validation_settigs import DEFAULT_VALIDATION, ValidationSettings
-from .exceptions import (
+from .exceptions.base import ExitError
+from .exceptions.fabrication import NoFactoryError
+from .exceptions.scopes import (
     ChildScopeNotFoundError,
-    ExitError,
     NoChildScopesError,
-    NoFactoryError,
     NoNonSkippedScopesError,
 )
 from .provider import BaseProvider
